***DO NOT USE BRANCH***
Testing out LGTM integration, which unfortunately needs specific files in the main/default branch of the repo. I'm going to remove this(/these) commit(s) from the revision history once I'm done trying it out, and will then either commit a final revision if I get it working and like it, but otherwise revert to the clean state before I tried it. So if you're unlucky enough to have pulled while I'm trying this, please do not make any modifications to this branch, as it's going to be deleted soon.